feat: create modules for js
